+++ packages/react/src/components/Popover/index.tsx
@@ -1,11 +1,6 @@
-export {
-  type AlignValue,
-  Content,
-  type ExtendedPlacement,
-  Root,
-  type PopoverContentProps,
-  type PopoverRootProps,
-  type PopoverTriggerProps,
-  type SideValue,
-  Trigger,
-} from './Popover'
+export { Root } from './Root'
+export { Content } from './Content'
+export { Trigger } from './Trigger'
+export type { PopoverRootProps } from './Root'
+export type { PopoverContentProps } from './Content'
+export type { PopoverTriggerProps } from './Trigger'
